requirements.txt
django
pillow
pytorch
torchvision
ultralytics